gui: hexview: add missing int conversion in AsciiText::render
There are three render() methods: OffsetText::render(),
AsciiText::render() and HexText::render(). In two of them the
division result for 'tot_lines' is correctly converted to int but
not in AsciiText::render(). During development Python 3 complained
about a float when expecting an int. So convert to int there as
well.

+++ gui/hexview.py
@@ -210,7 +210,7 @@ class AsciiText(BaseText):
         self.buffer.set_text('')
 
         bpl = self._parent.bpl
-        tot_lines = len(txt) / bpl
+        tot_lines = int(len(txt) / bpl)
 
         if len(txt) % bpl != 0:
             tot_lines += 1
